ml_api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
scikit-learn==1.3.0
numpy==1.25.2
pandas==2.1.0
tensorflow==2.12.0
transformers==4.33.2
torch==2.0.1
pytube==12.1.2
moviepy==1.0.3
python-dotenv==1.0.0
requests==2.28.2
pytest==7.4.2
httpx==0.24.1
joblib==1.3.2
sentence-transformers==2.2.2
